pyproject.toml
[tool.poetry]
name = "yless-pi-cowork-detection"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.84"
ultralytics = "^8.3.0"
netifaces = "^0.11.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
